omap3: Identify the CPU information in arch_cpu_init()
The function get_cpu_rev() is called multiple times
during execution resulting in probe into the IDCODE
register to extract the revision information.

This patch does the following:
- Moves the steps to identify static cpu information
  into arch_cpu_init().
- Updates configs for all omap3 boards to define
  CONFIG_ARCH_CPU_INIT.
- Updates get_cpu_rev() to return value calculated in
  arch_cpu_init().
- Since revision isn't expected to be longer than 8bits,
  get_cpu_rev() return a u8 value instead of u32.

+++ include/asm-arm/arch-omap3/sys_proto.h
@@ -42,7 +42,7 @@ void gpmc_init(void);
 void watchdog_init(void);
 void set_muxconf_regs(void);
 
-u32 get_cpu_rev(void);
+u8 get_cpu_rev(void);
 u32 get_mem_type(void);
 u32 get_sysboot_value(void);
 u32 is_gpmc_muxed(void);
